+++ modules/sagemaker/sagemaker-notebook/sagemaker_notebook/settings.py
@@ -42,6 +42,7 @@ class SeedFarmerParameters(CdkBaseSettings):
     role_arn: Optional[str] = Field(default=None)
     permissions_boundary_name: Optional[str] = Field(default=None)
     tags: Optional[Dict[str, str]] = Field(default=None)
+    custom_tags: Optional[Dict[str, str]] = Field(default=None)
 
     @field_validator("notebook_name")
     @classmethod