+++ modules/sagemaker/sagemaker-notebook/tests/test_app.py
@@ -23,3 +23,8 @@ def stack_defaults():
 
 def test_app(stack_defaults):
     import app  # noqa: F401
+
+
+def test_app_with_custom_tags(stack_defaults):
+    os.environ["SEEDFARMER_PARAMETER_CUSTOM_TAGS"] = '{"CustomerTag": "test-value"}'
+    import app  # noqa: F401