+++ modules/sagemaker/sagemaker-studio/app.py
@@ -73,6 +73,10 @@ if app_settings.parameters.tags:
     for tag_key, tag_value in app_settings.parameters.tags.items():
         aws_cdk.Tags.of(app).add(tag_key, tag_value)
 
+if app_settings.parameters.custom_tags:
+    for tag_key, tag_value in app_settings.parameters.custom_tags.items():
+        aws_cdk.Tags.of(app).add(tag_key, tag_value)
+
 aws_cdk.Tags.of(app).add("SeedFarmerDeploymentName", app_settings.settings.deployment_name)
 aws_cdk.Tags.of(app).add("SeedFarmerModuleName", app_settings.settings.module_name)
 aws_cdk.Tags.of(app).add("SeedFarmerProjectName", app_settings.settings.project_name)