+++ modules/sagemaker/sagemaker-studio/tests/test_app.py
@@ -58,3 +58,8 @@ def test_default_synthesizer(stack_defaults):
     os.environ["SEEDFARMER_PARAMETER_LOOKUP_ROLE_ARN"] = "dummy"
 
     import app as my_app  # noqa: F401
+
+
+def test_app_with_custom_tags(stack_defaults):
+    os.environ["SEEDFARMER_PARAMETER_CUSTOM_TAGS"] = '{"CustomerTag": "test-value"}'
+    import app  # noqa: F401